format scripts

+++ llm/vicuna-2/scripts/xformers_patch.py
@@ -12,7 +12,6 @@
 #    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
-
 """
 Directly copied the code from https://raw.githubusercontent.com/oobabooga/text-generation-webui/main/modules/llama_attn_hijack.py and made some adjustments
 """
@@ -28,7 +27,8 @@ from torch import nn
 try:
     import xformers.ops
 except ImportError:
-    logging.error("xformers not found! Please install it before trying to use it.")
+    logging.error(
+        "xformers not found! Please install it before trying to use it.")
 
 
 def replace_llama_attn_with_xformers_attn():
@@ -47,21 +47,15 @@ def xformers_forward(
     # pylint: disable=duplicate-code
     bsz, q_len, _ = hidden_states.size()
 
-    query_states = (
-        self.q_proj(hidden_states)
-        .view(bsz, q_len, self.num_heads, self.head_dim)
-        .transpose(1, 2)
-    )
-    key_states = (
-        self.k_proj(hidden_states)
-        .view(bsz, q_len, self.num_heads, self.head_dim)
-        .transpose(1, 2)
-    )
-    value_states = (
-        self.v_proj(hidden_states)
-        .view(bsz, q_len, self.num_heads, self.head_dim)
-        .transpose(1, 2)
-    )
+    query_states = (self.q_proj(hidden_states).view(bsz, q_len, self.num_heads,
+                                                    self.head_dim).transpose(
+                                                        1, 2))
+    key_states = (self.k_proj(hidden_states).view(bsz, q_len, self.num_heads,
+                                                  self.head_dim).transpose(
+                                                      1, 2))
+    value_states = (self.v_proj(hidden_states).view(bsz, q_len, self.num_heads,
+                                                    self.head_dim).transpose(
+                                                        1, 2))
 
     kv_seq_len = key_states.shape[-2]
     if past_key_value is not None:
@@ -71,8 +65,7 @@ def xformers_forward(
         query_states,
         key_states,
     ) = transformers.models.llama.modeling_llama.apply_rotary_pos_emb(
-        query_states, key_states, cos, sin, position_ids
-    )
+        query_states, key_states, cos, sin, position_ids)
     # [bsz, nh, t, hd]
 
     if past_key_value is not None:
@@ -93,8 +86,7 @@ def xformers_forward(
         if attention_mask is None or attention_mask[0, 0, 0, 1] == 0:
             # input and output should be of form (bsz, q_len, num_heads, head_dim)
             attn_output = xformers.ops.memory_efficient_attention(
-                query_states, key_states, value_states, attn_bias=None
-            )
+                query_states, key_states, value_states, attn_bias=None)
         else:
             # input and output should be of form (bsz, q_len, num_heads, head_dim)
             attn_output = xformers.ops.memory_efficient_attention(
@@ -105,15 +97,13 @@ def xformers_forward(
             )
         attn_weights = None
     else:
-        attn_weights = torch.matmul(
-            query_states, key_states.transpose(2, 3)
-        ) / math.sqrt(self.head_dim)
+        attn_weights = torch.matmul(query_states, key_states.transpose(
+            2, 3)) / math.sqrt(self.head_dim)
 
         if attn_weights.size() != (bsz, self.num_heads, q_len, kv_seq_len):
             raise ValueError(
                 f"Attention weights should be of size {(bsz * self.num_heads, q_len, kv_seq_len)}, but is"
-                f" {attn_weights.size()}"
-            )
+                f" {attn_weights.size()}")
 
         if attention_mask is not None:
             if attention_mask.size() != (bsz, 1, q_len, kv_seq_len):
@@ -122,20 +112,19 @@ def xformers_forward(
                 )
             attn_weights = attn_weights + attention_mask
             attn_weights = torch.max(
-                attn_weights, torch.tensor(torch.finfo(attn_weights.dtype).min)
-            )
+                attn_weights, torch.tensor(torch.finfo(attn_weights.dtype).min))
 
         # upcast attention to fp32
-        attn_weights = nn.functional.softmax(
-            attn_weights, dim=-1, dtype=torch.float32
-        ).to(query_states.dtype)
+        attn_weights = nn.functional.softmax(attn_weights,
+                                             dim=-1,
+                                             dtype=torch.float32).to(
+                                                 query_states.dtype)
         attn_output = torch.matmul(attn_weights, value_states)
 
         if attn_output.size() != (bsz, self.num_heads, q_len, self.head_dim):
             raise ValueError(
                 f"`attn_output` should be of size {(bsz, self.num_heads, q_len, self.head_dim)}, but is"
-                f" {attn_output.size()}"
-            )
+                f" {attn_output.size()}")
 
         attn_output = attn_output.transpose(1, 2)
